+++ NEWS.md
@@ -2,6 +2,7 @@
 # WVPlots 1.0.8 2019-01-26
 
  * Add baseline in lift curve plot.
+ * Truth targets.
 
 # WVPlots 1.0.7 2018-12-17